refactor vasp_interactive main code
