+++ vasp_interactive/vasp_interactive.py
@@ -3,6 +3,8 @@ Provides additional bug fix to
 https://gitlab.com/ase/ase/-/blob/master/ase/calculators/vasp/interactive.py
 May be merged with upstream by the end of day.
 """
+import psutil
+import signal
 from subprocess import Popen, PIPE
 from contextlib import contextmanager
 
@@ -14,75 +16,13 @@ from ase.io.vasp import write_vasp
 from ase.calculators.vasp.vasp import check_atoms
 from warnings import warn
 
-import time
-import os
-import sys
-import psutil
-import signal
-import re
-import numpy as np
-
-DEFAULT_KILL_TIMEOUT = 60
-
-
-class TimeoutException(Exception):
-    """Simple class for timeout"""
-
-    pass
-
-
-@contextmanager
-def time_limit(seconds):
-    """Usage:
-    try:
-        with time_limit(60):
-            do_something()
-    except TimeoutException:
-        raise
-    """
-
-    def signal_handler(signum, frame):
-        raise TimeoutException("Timed out closing VASP process.")
-
-    signal.signal(signal.SIGALRM, signal_handler)
-    signal.alarm(seconds)
-    try:
-        yield
-    finally:
-        signal.alarm(0)
-
-
-def _find_mpi_process(pid, mpi_program="mpirun", vasp_program="vasp_std"):
-    """Recursively search children processes with PID=pid and return the one
-    that mpirun (or synonyms) are the main command.
-    Note we currently do not support signal forwarding in srun and multiple node mpi
-    """
-    allowed_names = set(["mpirun", "mpiexec", "orterun", "oshrun", "shmemrun"])
-    allowed_vasp_names = set(["vasp_std", "vasp_gam", "vasp_ncl"])
-    if mpi_program:
-        allowed_names.add(mpi_program)
-    if vasp_program:
-        allowed_vasp_names.add(vasp_program)
-    process_list = [psutil.Process(pid)]
-    process_list.extend(process_list[0].children(recursive=True))
-    mpi_candidates = []
-    for proc in process_list:
-        # print(proc, proc.name())
-        if proc.name() in allowed_names:
-            # is the mpi process's direct children are vasp_std?
-            children = proc.children()
-            if len(children) > 0:
-                if children[0].name() in allowed_vasp_names:
-                    mpi_candidates.append(proc)
-    if len(mpi_candidates) > 1:
-        warn(
-            "More than 1 mpi processes are created. This may be a bug. I'll use the last one"
-        )
-    if len(mpi_candidates) > 0:
-        mpi_proc = mpi_candidates[-1]
-        return mpi_proc
-    else:
-        return None
+from .utils import DEFAULT_KILL_TIMEOUT, _int_version, time_limit, _find_mpi_process
+from .parse import (
+    parse_outcar_iterations,
+    parse_outcar_time,
+    parse_vaspout_energy,
+    parse_vaspout_forces,
+)
 
 
 class VaspInteractive(Vasp):
@@ -506,7 +446,9 @@ class VaspInteractive(Vasp):
         pid = self.process.pid
         mpi_process = _find_mpi_process(pid)
         if mpi_process is None:
-            warn("Cannot find the mpi process or you're using different ompi wrapper. Will not send stop signal to mpi.")
+            warn(
+                "Cannot find the mpi process or you're using different ompi wrapper. Will not send stop signal to mpi."
+            )
             return
         mpi_process.send_signal(sig)
         return
@@ -518,7 +460,9 @@ class VaspInteractive(Vasp):
         pid = self.process.pid
         mpi_process = _find_mpi_process(pid)
         if mpi_process is None:
-            warn("Cannot find the mpi process or you're using different ompi wrapper. Will not send continue signal to mpi.")
+            warn(
+                "Cannot find the mpi process or you're using different ompi wrapper. Will not send continue signal to mpi."
+            )
             return
         mpi_process.send_signal(sig)
         return
@@ -669,9 +613,7 @@ class VaspInteractive(Vasp):
                 # breakpoint()
                 self.results.update(dict(free_energy=energy_free, energy=energy_zero))
             except Exception as e:
-                raise RuntimeError((
-                    "Failed to obtain energy from calculator."
-                )) from e
+                raise RuntimeError(("Failed to obtain energy from calculator.")) from e
 
         if "forces" not in self.results.keys():
             try:
@@ -679,9 +621,7 @@ class VaspInteractive(Vasp):
                     dict(forces=self.read_forces(lines=outcar, vasp5=vasp5))
                 )
             except Exception as e:
-                raise RuntimeError((
-                    "Failed to obtain forces from calculator."
-                )) from e
+                raise RuntimeError(("Failed to obtain forces from calculator.")) from e
 
         if "magmom" not in self.results.keys():
             try:
@@ -892,127 +832,3 @@ class VaspInteractive(Vasp):
             )
         )
         return new
-
-
-# Following are functions parsing OUTCAR files which are not present in parent
-# Vasp calculator but can he helpful for job diagnosis
-
-
-def parse_outcar_iterations(lines):
-    """Read the whole iteration information (ionic + electronic) from OUTCAR lines"""
-    n_ion_scf = 0
-    n_elec_scf = []
-
-    for line in lines:
-        if "- Iteration" in line:
-            ni_, ne_ = list(map(int, re.findall(r"\d+", line)))
-            if ni_ > n_ion_scf:
-                n_ion_scf = ni_
-                n_elec_scf.append(ne_)
-            else:
-                n_elec_scf[ni_ - 1] = ne_
-    n_elec_scf = np.array(n_elec_scf)
-    return n_ion_scf, n_elec_scf
-
-
-def parse_outcar_time(lines):
-    """Parse the cpu and wall time from OUTCAR.
-    The mismatch between wall time and cpu time represents
-    the turn-around time in VaspInteractive
-
-    returns (cpu_time, wall_time)
-    if the calculation is not finished, both will be None
-    """
-    cpu_time = None
-    wall_time = None
-    for line in lines:
-        if "Total CPU time used (sec):" in line:
-            cpu_time = float(line.split(":")[1].strip())
-        if "Elapsed time (sec):" in line:
-            wall_time = float(line.split(":")[1].strip())
-    return cpu_time, wall_time
-
-
-def parse_vaspout_energy(vaspout, all=False):
-    """Parse the energy and force lines from vaspout or stdout
-    This should only be relevant when vasp5.x is involved, as a last resort
-    when both OUTCAR and vaspun.xml are truncated. Please use with care.
-
-    The vaspout format are almost identical to OSZICAR https://www.vasp.at/wiki/index.php/OSZICAR
-    FORCES:
-        (N x 3) fields of force
-    N F= XX E0= XX  d E =XX
-    The F & E0 line should most likely maintain the same format but there can be extra output lines like
-    vdW dispersion etc. Some codes can be traken from https://github.com/materialsproject/pymatgen/blob/v2022.9.8/pymatgen/io/vasp/outputs.py#L4253-L4394
-
-    For upstream methods in the calculator class, one must first check if the calc.txt is neither "-" nor None,
-    otherwise the vaspout lines cannot be parsed.
-    """
-    free_energies, zero_energies = [], []
-    ionic_pattern = re.compile(
-        (
-            r"(\d+)\s+F=\s*([\d\-\.E\+]+)\s+"
-            r"E0=\s*([\d\-\.E\+]+)\s+"
-            r"d\s*E\s*=\s*([\d\-\.E\+]+)$"
-        )
-    )
-    for line in vaspout:
-        line = line.strip()
-        if ionic_pattern.match(line.strip()):
-            m = ionic_pattern.match(line.strip())
-            free_energies.append(float(m.group(2)))
-            zero_energies.append(float(m.group(3)))
-    if all is False:
-        free_energies = free_energies[-1]
-        zero_energies = zero_energies[-1]
-    return free_energies, zero_energies
-
-
-def parse_vaspout_forces(vaspout, all=False):
-    """Parse the energy and force lines from vaspout or stdout
-    This should only be relevant when vasp5.x is involved, as a last resort
-    when both OUTCAR and vaspun.xml are truncated. Please use with care.
-
-    The vaspout format are almost identical to OSZICAR https://www.vasp.at/wiki/index.php/OSZICAR
-    FORCES:
-        (N x 3) fields of force
-    N F= XX E0= XX  d E =XX
-    The F & E0 line should most likely maintain the same format but there can be extra output lines like
-    vdW dispersion etc. Some codes can be taken from https://github.com/materialsproject/pymatgen/blob/v2022.9.8/pymatgen/io/vasp/outputs.py#L4253-L4394
-
-    For upstream methods in the calculator class, one must first check if the calc.txt is neither "-" nor None,
-    otherwise the vaspout lines cannot be parsed.
-    """
-    forces = []
-    current_line = 0
-    for i, line in enumerate(vaspout):
-        if i < current_line:
-            continue
-        else:
-            current_line = i
-        if "FORCES:" in line:
-            forces_this_step = []
-            while True:
-                current_line += 1
-                try:
-                    fx, fy, fz = np.fromstring(
-                        vaspout[current_line], dtype=float, sep=" "
-                    )
-                    forces_this_step.append([fx, fy, fz])
-                except Exception:
-                    break
-            forces.append(forces_this_step)
-    # forces = np.array(forces)
-    if all is False:
-        forces = np.array(forces[-1])
-    else:
-        # May be unusable if vasp.out is appended from previous calculation.
-        # Use all=True with caution
-        forces = np.array(forces)
-    return forces
-
-
-def _int_version(version_string):
-    """Get int version string"""
-    major = int(version_string.split(".")[0])
-    return major
